Tau energy scale implemented by new analyzers and gen matching in the before dilepton common sequence.
